stats19 file names for easy access

URL decoded file names. Currently there are 52 file names released by the DfT (Department for Transport) and the details include how these were obtained and would be kept up to date.

Format

A named list

Note

These were generated using the script in the data-raw directory (misc.Rmd file).

Examples

head(file_names)
#> [[1]]
#> [1] "dft-road-casualty-statistics-vehicle-provisional-mid-year-unvalidated-2025.csv"
#> 
#> [[2]]
#> [1] "dft-road-casualty-statistics-casualty-provisional-mid-year-unvalidated-2025.csv"
#> 
#> [[3]]
#> [1] "dft-road-casualty-statistics-collision-provisional-mid-year-unvalidated-2025.csv"
#> 
#> [[4]]
#> [1] "dft-road-casualty-statistics-casualty-2024.csv"
#> 
#> [[5]]
#> [1] "dft-road-casualty-statistics-vehicle-2024.csv"
#> 
#> [[6]]
#> [1] "dft-road-casualty-statistics-collision-2024.csv"
#>