Download and read-in severity adjustment factors

See the DfT's documentation on adjustment factors Annex: Update to severity adjustments methodology.

get_stats19_adjustments(
  data_dir = get_data_directory(),
  u = paste0("http://data.dft.gov.uk/road-accidents-safety-data/",
    "accident-and-casualty-adjustment-2004-to-2019.zip"),
  filename = "cas_adjustment_lookup_2019.csv",
  adj_folder = "adjustment-data"
)

Arguments

data_dir
Where sets of downloaded data would be found.

u
The URL of the zip file with adjustments to download

filename
The file name of the .csv file in the unzipped folder to read in

adj_folder
The folder name where R will look for the unzipped adjustment files

Details

See Estimating and adjusting for changes in the method of severity reporting for road accidents and casualty data: final report for details.

Examples

# \donttest{
adjustment = get_stats19_adjustments()

#> Unzipped files from DfT can be found in the folder:
#> /tmp/RtmpW9xaAb

#> accident-and-casualty-adjustment-2004-to-2019.zip
#> adjustment-data
#> DfTRoadSafety_Accidents_2009
#> DfTRoadSafety_Accidents_2009.zip
#> DfTRoadSafety_Accidents_2019
#> DfTRoadSafety_Accidents_2019.zip
#> DfTRoadSafety_Casualties_2009
#> DfTRoadSafety_Casualties_2009.zip
#> DfTRoadSafety_Casualties_2019
#> DfTRoadSafety_Casualties_2019.zip
#> dftRoadSafetyData_Accidents_2017
#> dftRoadSafetyData_Accidents_2017.zip
#> dftRoadSafetyData_Accidents_2018.csv
#> dftRoadSafetyData_Casualties_2017
#> dftRoadSafetyData_Casualties_2017.zip
#> dftRoadSafetyData_Vehicles_2017
#> dftRoadSafetyData_Vehicles_2017.zip
#> downlit
#> file119013bbcdf2
#> file119023119188
#> file11902f07e156
#> file11904ee4657a
#> file119072326fd
#> file119075347a63
#> file119075a659d9
#> Stats19-Data1979-2004
#> Stats19-Data1979-2004.zip

#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   accident_index = col_character(),
#>   Vehicle_Reference = col_double(),
#>   Casualty_Reference = col_double(),
#>   Adjusted_Serious = col_double(),
#>   Adjusted_Slight = col_double(),
#>   Injury_Based = col_double()
#> )

# }