Format column names of raw STATS19 data

This function takes messy column names and returns clean ones that work well with R by default. Names that are all lower case with no R-unfriendly characters such as spaces and - are returned.

format_column_names(column_names)

Arguments

column_names
Column names to be cleaned

Value

Column names cleaned.

Examples

# NOT RUN {
crashes_raw = read_accidents(year = 2017)
column_names = names(crashes_raw)
column_names
format_column_names(column_names = column_names)
# }